backend/requirements.txt
fastapi
uvicorn
pyyaml
simpleeval
sqlalchemy
psycopg2-binary
python-multipart
pydantic